+++ include/elf.hpp
@@ -7,6 +7,8 @@
 
 typedef std::pair<ELFIO::Elf64_Addr, ELFIO::Elf_Xword> X86FunctionMetadata;
 
+/// @brief Contains the raw code for an x86 function, which includes
+/// the bytes of the function, how big the function is, and the address
 struct X86Function {
     uint64_t address;
     uint64_t size;
